fix(couvoir): réparer la modale de lancement et la création d'incubation
Vérification du module Couvoir (incubations + incubateurs) — deux bugs
bloquants trouvés et corrigés :

- repro/index : l'attribut x-data de la modale « Nouveau lancement »
  injectait @json($incubationDurations) entre guillemets doubles, à
  l'intérieur d'un attribut lui-même délimité par des guillemets doubles.
  Les " du JSON fermaient l'attribut prématurément → tout le composant
  Alpine était cassé (openLaunch, isExternal, selectedProvider, duration
  « not defined »), donc la modale de création était totalement
  inutilisable. Le composant est déplacé dans une fonction
  incubationLaunchModal() (script) où le JSON vit dans un vrai contexte JS.

- StartIncubation : la durée issue du formulaire (chaîne « 21 ») était
  passée telle quelle à Carbon::addDays(), qui exige désormais un int
  → TypeError → 500 à CHAQUE création d'incubation. Cast en (int).

- routes : suppression des routes incubations.create et incubations.edit
  qui n'ont aucune méthode correspondante dans IncubationController
  (création/édition se font via la modale et les actions mirage/éclosion)
  → elles renvoyaient une erreur 500 si appelées.

Vérifié de bout en bout : la modale s'ouvre et bascule interne/externe,
une incubation se crée (œufs, durée, date d'éclosion calculée), la
machine passe à « Occupé », et la page affiche le cycle actif. Aucun
débordement latéral à 390px ni 1440px.

+++ app/Actions/Incubation/StartIncubation.php
@@ -17,7 +17,7 @@ class StartIncubation
             $incubator = Incubator::findOrFail($data['incubator_id']);
             $batchId = $this->resolveBatchId($data);
             
-            $duration = $data['duration'] ?? 21; // Espèce par défaut : Poule
+            $duration = (int) ($data['duration'] ?? 21); // Espèce par défaut : Poule
 
             $incubation = Incubation::create([
                 'batch_id'            => $batchId,
